+++ server/main/src/server/open_file.rs
@@ -10,20 +10,13 @@ impl MinecraftLanguageServer {
         let mut temp_files = server_data.temp_files.borrow_mut();
         let temp_lint = server_data.temp_lint.borrow();
 
-        let diagnostics = if let Some(workspace_file) = workspace_files.get(&file_path) {
-            let shader_files = workspace_file.parent_shaders().borrow();
-            let mut update_list = HashMap::new();
-            shader_files.iter().for_each(|(shader_path, shader_file)| {
-                self.lint_workspace_shader(shader_file, shader_path, &mut update_list);
-            });
-
-            self.collect_diagnostics(&update_list)
+        if workspace_files.contains_key(&file_path) {
+            None
         } else {
             let temp_file = TempFile::new(&mut parser, &file_path, params.text_document.text);
             let diagnostics = self.lint_temp_file(&temp_file, &file_path, params.text_document.uri, *temp_lint);
             temp_files.insert(file_path, temp_file);
-            diagnostics
-        };
-        Some(diagnostics)
+            Some(diagnostics)
+        }
     }
 }